+++ src/main/java/org/dpsoftware/utilities/CommonUtility.java
@@ -402,11 +402,12 @@ public class CommonUtility {
      * Turn ON LEDs when Luciferin starts
      */
     public static void turnOnLEDs() {
-        if (!Constants.Effect.BIAS_LIGHT.equals(LocalizedEnum.fromBaseStr(Constants.Effect.class, FireflyLuciferin.config.getEffect()))
-                && !Constants.Effect.MUSIC_MODE_VU_METER.equals(LocalizedEnum.fromBaseStr(Constants.Effect.class, FireflyLuciferin.config.getEffect()))
-                && !Constants.Effect.MUSIC_MODE_VU_METER_DUAL.equals(LocalizedEnum.fromBaseStr(Constants.Effect.class, FireflyLuciferin.config.getEffect()))
-                && !Constants.Effect.MUSIC_MODE_BRIGHT.equals(LocalizedEnum.fromBaseStr(Constants.Effect.class, FireflyLuciferin.config.getEffect()))
-                && !Constants.Effect.MUSIC_MODE_RAINBOW.equals(LocalizedEnum.fromBaseStr(Constants.Effect.class, FireflyLuciferin.config.getEffect()))) {
+        Constants.Effect effectInUse = LocalizedEnum.fromBaseStr(Constants.Effect.class, FireflyLuciferin.config.getEffect());
+        if (!Constants.Effect.BIAS_LIGHT.equals(effectInUse)
+                && !Constants.Effect.MUSIC_MODE_VU_METER.equals(effectInUse)
+                && !Constants.Effect.MUSIC_MODE_VU_METER_DUAL.equals(effectInUse)
+                && !Constants.Effect.MUSIC_MODE_BRIGHT.equals(effectInUse)
+                && !Constants.Effect.MUSIC_MODE_RAINBOW.equals(effectInUse)) {
             if (FireflyLuciferin.config.isToggleLed()) {
                 if (FireflyLuciferin.config.isWifiEnable()) {
                     String[] color = FireflyLuciferin.config.getColorChooser().split(",");
